Move GameFrame into client ui
