+++ engine/src/main/kotlin/world/gregs/voidps/engine/client/instruction/handle/ScreenChangeHandler.kt
@@ -2,8 +2,8 @@ package world.gregs.voidps.engine.client.instruction.handle
 
 import world.gregs.voidps.engine.client.instruction.InstructionHandler
 import world.gregs.voidps.engine.client.ui.isOpen
+import world.gregs.voidps.engine.client.ui.setDisplayMode
 import world.gregs.voidps.engine.entity.character.player.Player
-import world.gregs.voidps.engine.entity.character.player.setDisplayMode
 import world.gregs.voidps.network.instruct.ChangeDisplayMode
 
 class ScreenChangeHandler : InstructionHandler<ChangeDisplayMode>() {
