+++ engine/src/test/kotlin/world/gregs/voidps/engine/client/ui/InterfaceTest.kt
@@ -5,7 +5,6 @@ import io.mockk.mockkStatic
 import io.mockk.spyk
 import org.junit.jupiter.api.BeforeEach
 import org.koin.test.mock.declare
-import world.gregs.voidps.engine.entity.character.player.PlayerGameFrame
 import world.gregs.voidps.engine.entity.definition.InterfaceDefinitions
 import world.gregs.voidps.engine.event.Events
 import world.gregs.voidps.engine.script.KoinMock
@@ -18,14 +17,14 @@ abstract class InterfaceTest : KoinMock() {
     internal lateinit var interfaces: Interfaces
     internal lateinit var open: MutableSet<String>
     internal lateinit var definitions: InterfaceDefinitions
-    internal lateinit var gameframe: PlayerGameFrame
+    internal lateinit var gameframe: GameFrame
 
     @BeforeEach
     open fun setup() {
         client = mockk(relaxed = true)
         events = mockk(relaxed = true)
         definitions = declare { mockk(relaxed = true) }
-        gameframe = spyk(PlayerGameFrame())
+        gameframe = spyk(GameFrame())
         open = mutableSetOf()
         interfaces = spyk(Interfaces(events, client, definitions, gameframe, open))
         mockkStatic("world.gregs.voidps.network.encode.InterfaceEncodersKt")
